requirements.txt
discord.py[voice]>=1.2.5
youtube-dl>=2017.12.31